pyproject.toml
[tool.poetry]
name = "cat-in-the-dat-ii"
version = "0.1.0"
description = ""
authors = ["tenajima <tenajima@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pandas = "^1.0.1"
lightgbm = "^2.3.1"
scikit-learn = "^0.22.1"
luigi = "^2.8.11"
gokart = "^0.3.11"
python-dotenv = "^0.11.0"
category-encoders = "^2.1.0"
optuna = "^1.1.0"
nyaggle = {git = "https://github.com/nyanp/nyaggle.git", rev = "master"}
mlflow = "^1.6.0"
click = "^7.0"

[tool.poetry.dev-dependencies]
jupyterlab = "^1.2.6"
pytest = "^5.3.5"
black = "^19.10b0"
flake8 = "^3.7.9"
jupyterlab_code_formatter = "^1.1.0"
thunderbolt = "^0.0.22"
kaggle = "^1.5.6"
isort = "^4.3.21"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
